Split programs/server into actual program and library (#11186)
* Split programs/server into actual program and library
