+++ programs/server/ya.make
@@ -11,19 +11,8 @@ PEERDIR(
 SRCS(
     clickhouse-server.cpp
 
-    HTTPHandler.cpp
-    HTTPHandlerFactory.cpp
-    InterserverIOHTTPHandler.cpp
     MetricsTransmitter.cpp
-    MySQLHandler.cpp
-    MySQLHandlerFactory.cpp
-    NotFoundHandler.cpp
-    PrometheusMetricsWriter.cpp
-    PrometheusRequestHandler.cpp
-    ReplicasStatusHandler.cpp
-    StaticRequestHandler.cpp
     Server.cpp
-    TCPHandler.cpp
 )
 
 END()
