+++ src/Server/HTTPHandlerFactory.cpp
@@ -4,6 +4,7 @@
 #include <re2/stringpiece.h>
 #include <common/find_symbols.h>
 #include <Poco/StringTokenizer.h>
+#include <Poco/Util/LayeredConfiguration.h>
 
 #include "HTTPHandler.h"
 #include "NotFoundHandler.h"
